add a changes to localization
