+++ package.json
@@ -11,6 +11,7 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "husky": "^9.0.11"
+    "husky": "^9.0.11",
+    "openai": "^4.38.3"
   }
 }
